Profile UI in delivery and owner

+++ owner/src/pages/OwnerProfile.js
@@ -6,6 +6,7 @@ import UpdateProfileDetails from "../components/UpdateProfileDetails";
 import axios from "axios";
 import { ownerLogIn, setCurrentColor, userData } from "../redux/user/userSlice";
 import { useDispatch } from "react-redux";
+import swal from "sweetalert";
 
 const OwnerProfile = () => {
   const dispatch = useDispatch();
@@ -17,7 +18,7 @@ const OwnerProfile = () => {
   const [isLoading, setIsLoading] = useState(null);
   const [order, setOrder] = useState([]);
   const user = useSelector((state) => state.userData.user);
-  console.log("===user",user);
+  console.log("===user", user);
   useEffect(() => {
     (async () => {
       if (user != null) {
@@ -31,15 +32,26 @@ const OwnerProfile = () => {
         setIsLoading(false);
       }
     })();
-    dispatch(setCurrentColor("slate-800"))
+    dispatch(setCurrentColor("slate-800"));
   }, []);
   const history = useNavigate();
 
   const handleLogOut = () => {
-    localStorage.clear();
-    dispatch(ownerLogIn(false));
-    dispatch(userData(null));
-    history("/");
+    swal({
+      title: "Are you Sure! you want to logout?",
+      icon: "warning",
+      buttons: ["NO", "YES"],
+      cancelButtonColor: "#DD6B55",
+      confirmButtonColor: "#DD6B55",
+      dangerMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        localStorage.clear();
+        dispatch(ownerLogIn(false));
+        dispatch(userData(null));
+        history("/");
+      }
+    });
   };
 
   return (
@@ -63,137 +75,173 @@ const OwnerProfile = () => {
               Remove
             </button>
           </div> */}
-          <div className="border-b-2 border-black/30"></div>
+          <div className="border-b-2 border-black/30 mt-5"></div>
         </div>
 
         <div className="flex flex-wrap">
           <div className="w-full sm:w-1/5 p-5">
             <ul className="space-y-3">
               <li
-                className="text-lg border-b-2 cursor-pointer"
+                className={`text-lg ${
+                  openTab == 1
+                    ? "border-slate-700 border-2"
+                    : "bg-slate-200 border-slate-200 border-2"
+                } font-mono font-semibold text-black pl-5 py-2 cursor-pointer  rounded-3xl hover:pl-8 duration-300`}
                 onClick={() => setOpenTab(1)}
               >
                 Profile
               </li>
-              <li
-                className="text-lg border-b-2 cursor-pointer"
-                onClick={() => setOpenTab(2)}
-              >
-                Your Restaurant{" "}
-              </li>
-              <li className="text-lg border-b-2 cursor-pointer">
-                <button
-                  onClick={handleLogOut}
-                  className="inline-block bg-inherit hover:text-white hover:bg-red-600 font-bold  rounded  px-4  py-[6px] text-xs uppercase  text-red-600 transition hover:scale-110 hover:shadow-xl focus:outline-none focus:ring active:text-red-500"
+              {user?.isApproved != "Not Request" && (
+                <li
+                  className={`text-lg ${
+                    openTab == 2
+                      ? "border-slate-700 border-2"
+                      : "bg-slate-200 border-slate-200 border-2"
+                  } font-mono font-semibold text-black pl-5 py-2 cursor-pointer  rounded-3xl hover:pl-8 duration-300`}
+                  onClick={() => setOpenTab(2)}
                 >
-                  Logout
-                </button>
+                  Your Restaurant{" "}
+                </li>
+              )}
+
+              <li className="bg-black h-1 w-full rounded-full"></li>
+              <li
+                className={`text-base ${"bg-slate-100 border-slate-200 border-2"} font-mono font-semibold text-black pl-5 py-2 cursor-pointer  rounded-3xl hover:pl-8 duration-300 hover:border-black hover:text-white hover:bg-black`}
+                onClick={() => setupdateProfile(true)}
+              >
+                <i className="fas fa-repeat"></i> Update Profile
+              </li>
+              <li
+                className={`text-base ${"bg-slate-100 border-slate-200 border-2"} font-mono font-semibold text-black pl-5 py-2 cursor-pointer  rounded-3xl hover:pl-8 duration-300 hover:border-black hover:text-white hover:bg-black`}
+                onClick={() => setChangePassword(true)}
+              >
+                Change Password
+              </li>
+              <li
+                className={`text-base ${"bg-slate-100 border-slate-200 border-2"} font-mono font-semibold text-black pl-5 py-2 cursor-pointer  rounded-3xl hover:pl-8 duration-300 hover:border-black hover:text-white hover:bg-black`}
+                onClick={handleLogOut}
+              >
+                LogOut
               </li>
             </ul>
           </div>
 
           {/* profile module  */}
           <div
-            className={`${openTab === 1 ? "block" : "hidden"
-              } w-full sm:w-4/5 p-5`}
+            className={`${
+              openTab === 1 ? "block" : "hidden"
+            } w-full sm:w-4/5 p-5`}
           >
-            <div>
-              <h1 className="text-xl font-semibold pb-5 capitalize">
-                your profile
+            <div className="mb-2 shadow-md rounded-tl-3xl rounded-br-3xl bg-slate-700 bg-opacity-40 ">
+              <h1 className="text-2xl font-normal text-white capitalize border-b-4 border-yellow-300 p-5">
+                Your profile
               </h1>
-              <div className="flex gap-7 ">
-                <div>
-                  <ul className="">
-                    <li className="py-3 text-lg font-extralight">Name</li>
-                    <li className=" text-lg font-extralight">Number</li>
-                    <li className="py-3 text-lg font-extralight">Email</li>
-                  </ul>
-                </div>
-                <div>
-                  <ul className="border-l-2 pl-5">
-                    <li className="py-3 text-lg font-normal capitalize">
-                      {user != null && user.name}
-                    </li>
-                    <li className=" text-lg font-normal">{user != null && user.number}</li>
-                    <li className="py-3 text-lg font-normal">{user != null && user.email}</li>
-                    <li className="py-3 text-lg font-normal">
-                      <button
-                        type="button"
-                        onClick={() => setChangePassword(true)}
-                        id="password"
-                        value="Change Password"
-                        className="inline-block mr-2 bg-white hover:text-white hover:bg-blue-400 -bottom-4 font-bold  rounded border border-current px-4 py-[6px] text-xs uppercase  text-blue-400 transition hover:scale-110 hover:shadow-xl focus:outline-none focus:ring active:text-blue-400"
-                      >
-                        <i className="fas fa-repeat"></i> Change Password
-                      </button>
-                      <button
-                        onClick={() => setupdateProfile(true)}
-                        className="inline-block bg-white hover:text-white hover:bg-blue-600 -bottom-4 font-bold  rounded border border-current px-8 py-[6px] text-xs uppercase  text-blue-600 transition hover:scale-110 hover:shadow-xl focus:outline-none focus:ring active:text-blue-500"
-                      >
-                        Update Profile{" "}
-                      </button>
-                    </li>
-                  </ul>
-                </div>
+              <div className="bg-slate-200 pl-5 rounded-br-3xl">
+                <table className="table-auto border-spacing-y-3 border-separate">
+                  <tr className="">
+                    <td className="text-slate-700 text-lg text-semibold pr-5 w-1/6">
+                      Name
+                    </td>
+                    <td className="text-slate-500 font-semibold capitalize bg-slate-50 w-full md:w-5/6 bg-opacity-20 p-2 rounded">
+                      {user != null ? user?.name : ""}
+                    </td>
+                  </tr>
+                  <tr>
+                    <td className="text-slate-700 text-lg text-semibold pr-5 ">
+                      Mobile No.
+                    </td>
+                    <td className="text-slate-500 font-semibold capitalize bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                      {user != null ? user?.number : ""}
+                    </td>
+                  </tr>
+                  <tr>
+                    <td className="text-slate-700 text-lg text-semibold pr-5">
+                      E-mail
+                    </td>
+                    <td className="text-slate-500 font-semibold bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                      {user != null ? user?.email : ""}
+                    </td>
+                  </tr>
+                </table>
               </div>
             </div>
-            <div className="flex flex-wrap">
-              <div className="flex w-full sm:w-1/2 flex-col">
-                <h1 className="text-lg font-semibold py-5">Bank details</h1>
-                <div className="flex">
-                  <ul className="pr-2">
-                    <li className="py-3 text-lg font-extralight">
-                      Account No.
-                    </li>
-                    <li className=" text-lg font-extralight">IFSC Code</li>
-                    <li className="py-3 text-lg font-extralight">
-                      Account type
-                    </li>
-                  </ul>
-                  <ul className="border-l-2 pl-5">
-                    <li className="py-3 text-lg font-normal capitalize">
-                      {user != null && user?.bankDetails?.ACnumber}
-                    </li>
-                    <li className=" text-lg font-normal">
-                      {user != null && user?.bankDetails?.IFSC}
-                    </li>
-                    <li className="py-3 text-lg font-normal">
-                      {user != null && user?.bankDetails?.actype}
-                    </li>
-                  </ul>
+            {user?.isApproved != "Not Request" && (
+              <div className="flex flex-wrap justify-between mt-5">
+                <div className="md:w-[48%] mb-2 shadow-md rounded-tl-3xl rounded-br-3xl bg-slate-700 bg-opacity-40 ">
+                  <h1 className="text-2xl font-normal text-white capitalize border-b-4 border-yellow-300 p-5">
+                    Bank details
+                  </h1>
+                  <div className="bg-slate-200 pl-5 rounded-br-3xl">
+                    <table className="table-auto border-spacing-y-3 border-separate">
+                      <tr className="">
+                        <td className="text-slate-700 text-lg text-semibold pr-5 w-1/6">
+                          Account No.
+                        </td>
+                        <td className="text-slate-500 font-semibold capitalize bg-slate-50 w-full md:w-5/6 bg-opacity-20 p-2 rounded">
+                          {user != null && user?.bankDetails?.ACnumber}
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-slate-700 text-lg text-semibold pr-5 ">
+                          IFSC Code
+                        </td>
+                        <td className="text-slate-500 font-semibold capitalize bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                          {user != null && user?.bankDetails?.IFSC}
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-slate-700 text-lg text-semibold pr-5">
+                          Account type
+                        </td>
+                        <td className="text-slate-500 font-semibold bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                          {user != null && user?.bankDetails?.actype}
+                        </td>
+                      </tr>
+                    </table>
+                  </div>
+                </div>
+                <div className="md:w-[48%] mb-2 shadow-md rounded-tl-3xl rounded-br-3xl bg-slate-700 bg-opacity-40 ">
+                  <h1 className="text-2xl font-normal text-white capitalize border-b-4 border-yellow-300 p-5">
+                    PAN Card details
+                  </h1>
+                  <div className="bg-slate-200 pl-5 rounded-br-3xl">
+                    <table className="table-auto border-spacing-y-3 border-separate">
+                      <tr className="">
+                        <td className="text-slate-700 text-lg text-semibold pr-5 w-1/6">
+                          Holder Name
+                        </td>
+                        <td className="text-slate-500 font-semibold capitalize bg-slate-50 w-full md:w-5/6 bg-opacity-20 p-2 rounded">
+                          {user != null && user?.panCard?.holderName}
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-slate-700 text-lg text-semibold pr-5 ">
+                          PAN No.
+                        </td>
+                        <td className="text-slate-500 font-semibold capitalize bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                          {user != null && user?.panCard?.number}
+                        </td>
+                      </tr>
+                      <tr>
+                        <td className="text-slate-700 text-lg text-semibold pr-5 ">
+                          PAN Status
+                        </td>
+                        <td className="text-slate-500 font-semibold capitalize bg-slate-50  w-full md:w-5/6 bg-opacity-20 p-2 rounded ">
+                          Verified
+                        </td>
+                      </tr>
+                    </table>
+                  </div>
                 </div>
               </div>
-              <div className="flex w-full sm:w-1/2 flex-col">
-                <h1 className="text-lg font-semibold py-5">PAN Card details</h1>
-                <div className="flex">
-                  <ul className="pr-2">
-                    <li className="py-3 text-lg font-extralight">Name</li>
-                    <li className="py-3 text-lg font-extralight">PAN No.</li>
-                  </ul>
-                  <ul className="border-l-2 pl-5">
-                    <li className="py-3 text-lg font-normal capitalize">
-                      {user != null && user?.panCard?.holderName}
-                    </li>
-                    <li className="py-3 text-lg font-normal">
-                      {user != null && user?.panCard?.number}
-                    </li>
-                  </ul>
-                </div>
-              </div>
-            </div>
-            {/* <Link
-              to="/"
-              className={`px-2 flex py-2 flex-1 justify-center items-center font-semibold text-xl text-white rounded-md shadow-lg hover:pr-10 duration-500 bg-blue-400 w-36`}
-              onClick={handleLogOut}
-            >
-              Log out
-            </Link> */}
+            )}
           </div>
 
           {/* your restaurant  */}
           <div
-            className={`${openTab === 2 ? "block" : "hidden"
-              } w-full sm:w-4/5 p-5`}
+            className={`${
+              openTab === 2 ? "block" : "hidden"
+            } w-full sm:w-4/5 p-5`}
           >
             <h1 className="text-xl font-semibold pb-5 capitalize">
               Your Restaurant
@@ -214,8 +262,16 @@ const OwnerProfile = () => {
                     <li className="py-3 text-lg font-normal capitalize">
                       {user != null && user.name}
                     </li>
-                    <li className=" text-lg font-normal">{user != null && user.email}</li>
-                    <li className="py-3 text-lg font-normal">{`${user != null && user.address?.street} ${user != null && user.address?.area} ${user != null && user.address?.city} ${user != null && user.address?.state}-${user != null && user.address?.pincode}`}</li>
+                    <li className=" text-lg font-normal">
+                      {user != null && user.email}
+                    </li>
+                    <li className="py-3 text-lg font-normal">{`${
+                      user != null && user.address?.street
+                    } ${user != null && user.address?.area} ${
+                      user != null && user.address?.city
+                    } ${user != null && user.address?.state}-${
+                      user != null && user.address?.pincode
+                    }`}</li>
                   </ul>
                 </div>
               </div>
